refactor(primitives): Extract module for BenchmarkData
